python/config.example.py
# file da mettere dentro la cartella /instance (in teoria non va cambiato nulla)
import os
from dotenv import load_dotenv

load_dotenv()

SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL')
SQLALCHEMY_TRACK_MODIFICATIONS = False