python/requirements.txt
attrs==26.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.3
cryptography==48.0.0
flasgger==0.9.7.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.3
flask-marshmallow==1.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.5.0
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
marshmallow==4.3.0
marshmallow-sqlalchemy==1.5.0
mistune==3.2.1
more-itertools==11.0.2
oracledb==4.0.0
packaging==26.2
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
sqlacodegen==4.0.3
SQLAlchemy==2.0.49
typeguard==4.5.1
typing_extensions==4.15.0
Werkzeug==3.1.8
wheel==0.46.3
